Move extensions to md_extensions
