+++ rdm/init_files/data/system.yml
@@ -3,9 +3,9 @@ project_name: PROJECT
 
 # List optional extensions
 extension_load_list:
-  - 'rdm.audit_notes.AuditNoteExtension'
-  - 'rdm.section_numbers.SectionNumberExtension'
-  - 'rdm.vocabulary_extension.VocabularyExtension'
+  - 'rdm.md_extensions.audit_notes.AuditNoteExtension'
+  - 'rdm.md_extensions.section_numbers.SectionNumberExtension'
+  - 'rdm.md_extensions.vocabulary_extension.VocabularyExtension'
 
 # Standard Version
 standard: IEC62304:2006
